solving int and size_t compare problem v2

+++ malashenko.dmitrii/P3/main.cpp
@@ -63,7 +63,7 @@ namespace malasenko {
     size_t total = rows * cols;
     size_t i = rows - 1, j = 0;
     size_t step = 0, cnt = 0;
-    int top = 0,  botom = rows - 1, left = 0, right = cols - 1;
+    size_t top = 0,  botom = rows - 1, left = 0, right = cols - 1;
     while (cnt != total) {
       while (i > top && cnt < total) {
         mtx[i * cols + j] -= ++step;
